third_party/WebKit/LayoutTests/external/wpt/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-candidate-insert-before-expected.txt

This is a testharness.js-based test.
FAIL inserting another source before the candidate assert_equals: expected "#a" but got "#b"
Harness: the test ran to completion.
